+++ src/Braintree/TransactionRequest.cs
@@ -45,6 +45,7 @@ namespace Braintree
         [ObsoleteAttribute("use DeviceData instead", false)]
         public string FraudMerchantId { get; set; }
         public string Channel { get; set; }
+        public string ExchangeRateQuoteId { get; set; }
         public string OrderId { get; set; }
         public string ProductSku { get; set; }
         [ObsoleteAttribute("use TransactionSource instead", false)]
@@ -137,6 +138,7 @@ namespace Braintree
             builder.AddElement("order-id", OrderId);
             builder.AddElement("product-sku", ProductSku);
             builder.AddElement("channel", Channel);
+            builder.AddElement("exchange-rate-quote-id",ExchangeRateQuoteId);
 // Remove this pragma warning when we remove DeviceSessionId, FraudMerchantId, and Recurring.
 // We have this so we can build the SDK without obsolete error messages
 #pragma warning disable 618